ir: add template usage check, make basename global
A model may contain templates, but not use them. Create the
check_templates global functions that, given a row container
object, checks whether the model it represents actually makes
use of template components. This allows for Isograph exporter
to determine whether it should rename the components or not.
It may also be useful for other exporters as well in the
future.

Signed-off-by: Ergys Dona <errikosd@gmail.com>